+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #36: Select node with cursor
 - Issue #37: Select edge with cursor
 - Issue #46: Zoom buttons
+- Issue #55: Delete edge by keyboard and detail button
 
 ## [4.2.0] - 2019-12-08